+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlapi/fakeevents/IntegrationTestWithFakeEvents.java
@@ -1,4 +1,4 @@
 package uk.gov.hmcts.reform.sandl.snlapi.fakeevents;
 
-public interface IntegrationTestWithFakeBackend {
+public interface IntegrationTestWithFakeEvents {
 }
