+++ src/test/integration/uk/gov/hmcts/reform/sandl/snlapi/realevents/IntegrationTestWithRealEvents.java
@@ -1,4 +1,4 @@
 package uk.gov.hmcts.reform.sandl.snlapi.realevents;
 
-public interface IntegrationTestWithRealBackend {
+public interface IntegrationTestWithRealEvents {
 }
